Sync compiles, still need more work on the model system

+++ Core/src/gfx/Material.h
@@ -17,7 +17,7 @@ namespace tryn::gfx
 	public:
 		Material(IGraphics& gfx, const aiMaterial& material, const std::filesystem::path& path);
 		VertexBuffer ExtractVertices(const aiMesh& mesh) const noexcept;
-		std::vector<unsigned short> ExtractIndices(const aiMesh& mesh) const noexcept;
+		std::vector<int> ExtractIndices(const aiMesh& mesh) const noexcept;
 		std::shared_ptr<IVertexBuffer> MakeVertexBindable(IGraphics& gfx, const aiMesh& mesh, glm::vec3 scale = {1.0f,1.0f,1.0f}) const;
 		std::shared_ptr<IIndexBuffer> MakeIndexBindable(IGraphics& gfx, const aiMesh& mesh) const;
 		std::vector<Technique> GetTechniques() const noexcept;
